reference #279, needs testing but this might be all it takes

+++ helpers/checks/torprebypass.js
@@ -12,37 +12,40 @@ const { Bypass } = require(__dirname+'/../../db/')
 
 module.exports = async (req, res, next) => {
 
-	//early byapss is only needed for tor users
+	//early bypass is only needed for tor users
 	if (!res.locals.tor) {
 		return next();
 	}
 
-	const input = req.body.captcha;
-	const captchaId = req.cookies.captchaid;
 	let bypassId = req.signedCookies.bypassid;
-	if (input && !bypassId) {
-		// try to get the captcha from the DB
-		try {
-			await checkCaptcha(input, captchaId);
-		} catch (err) {
-			deleteTempFiles(req).catch(e => console.error);
-			if (err instanceof Error) {
-				return next(err);
+
+	if (blockBypass.enabled) {
+		const input = req.body.captcha;
+		const captchaId = req.cookies.captchaid;
+		if (input && !bypassId) {
+			// try to get the captcha from the DB
+			try {
+				await checkCaptcha(input, captchaId);
+			} catch (err) {
+				deleteTempFiles(req).catch(e => console.error);
+				if (err instanceof Error) {
+					return next(err);
+				}
+				const page = (req.body.minimal || req.path === '/blockbypass' ? 'bypass' : 'message');
+				return dynamicResponse(req, res, 403, page, {
+					'title': 'Forbidden',
+					'message': err,
+					'redirect': req.headers.referer,
+				});
 			}
-			const page = (req.body.minimal || req.path === '/blockbypass' ? 'bypass' : 'message');
-			return dynamicResponse(req, res, 403, page, {
-				'title': 'Forbidden',
-				'message': err,
-				'redirect': req.headers.referer,
-			});
+			res.locals.solvedCaptcha = true;
+			res.clearCookie('captchaid');
+			remove(`${uploadDirectory}/captcha/${captchaId}.jpg`).catch(e => { console.error(e) });
 		}
-		res.locals.solvedCaptcha = true;
-		res.clearCookie('captchaid');
-		remove(`${uploadDirectory}/captcha/${captchaId}.jpg`).catch(e => { console.error(e) });
 	}
 
-	if (res.locals.solvedCaptcha) {
-		//they dont have a valid bypass, but just solved a captcha, so give them a new one
+	if (res.locals.solvedCaptcha || !blockBypass.enabled) {
+		//give bypass ("ip") to tor user if they solved captcha or blockbypass is disabled
 		const newBypass = await Bypass.getBypass();
 		const newBypassId = newBypass.insertedId;
 		bypassId = newBypassId.toString();
